Revert "Added mappers for each ORM and worked on the first exercise (d), creating a Cliente_Particular. Other fixes and small changes."
This reverts commit 1b0fdb202059b83c33c9a7dec66e2e3c950f18ce.

+++ F2/src/main/java/isel/sisinf/grp02/data_mappers/ICliente_ParticularMapper.java
@@ -2,5 +2,5 @@ package isel.sisinf.grp02.data_mappers;
 
 import isel.sisinf.grp02.orm.Cliente_Particular;
 
-public interface ICliente_ParticularMapper extends IDataMapper<Cliente_Particular, Integer> {
+public interface ICliente_ParticularMapper extends IDataMapper<Cliente_Particular, Long> {
 }
